tests/tools/visualization_requirements.txt
# 測試結果可視化分析工具依賴
# Test Result Visualization Analysis Tool Dependencies

# 核心數據處理
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# 機器學習和統計分析
scikit-learn>=1.1.0
statsmodels>=0.13.0

# 圖表和可視化
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web框架和儀表板
flask>=2.2.0
flask-socketio>=5.2.0
fastapi>=0.85.0
uvicorn>=0.18.0
jinja2>=3.1.0

# PDF報告生成
weasyprint>=57.0
reportlab>=3.6.0

# 數據庫
sqlite3  # 內建於Python

# HTML模板和CSS
jinja2>=3.1.0
cssutils>=2.5.0

# 配置管理
pyyaml>=6.0
toml>=0.10.0

# 文件處理
lxml>=4.9.0
beautifulsoup4>=4.11.0

# 日期時間處理
python-dateutil>=2.8.0

# 並發和異步
asyncio  # 內建於Python
threading  # 內建於Python

# 網絡和HTTP
requests>=2.28.0
aiohttp>=3.8.0

# 開發和測試工具（可選）
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# 性能分析工具（可選）
memory-profiler>=0.60.0
psutil>=5.9.0

# 額外的圖表庫（可選）
bokeh>=2.4.0
altair>=4.2.0

# 數據導出格式支持
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# 圖像處理（用於圖表導出）
pillow>=9.0.0
kaleido>=0.2.1  # plotly圖表導出

# 命令行界面
click>=8.0.0
argparse  # 內建於Python

# 日誌記錄
logging  # 內建於Python
colorlog>=6.7.0

# 類型提示
typing-extensions>=4.3.0

# 緩存
diskcache>=5.4.0

# 壓縮和歸檔
zipfile  # 內建於Python
tarfile  # 內建於Python

# 操作系統接口
pathlib  # 內建於Python
os  # 內建於Python
shutil  # 內建於Python

# JSON和序列化
json  # 內建於Python
pickle  # 內建於Python

# 正則表達式
re  # 內建於Python

# 數學和統計
math  # 內建於Python
statistics  # 內建於Python

# 多進程和線程
multiprocessing  # 內建於Python
concurrent.futures  # 內建於Python

# 網絡和套接字
socket  # 內建於Python
urllib  # 內建於Python

# 加密和哈希
hashlib  # 內建於Python
secrets  # 內建於Python

# 時間和時區
time  # 內建於Python
datetime  # 內建於Python
calendar  # 內建於Python

# 環境變量
environ  # 通過os.environ訪問

# 警告控制
warnings  # 內建於Python

# 垃圾回收
gc  # 內建於Python

# 系統信息
platform  # 內建於Python
sys  # 內建於Python

# 迭代工具
itertools  # 內建於Python
collections  # 內建於Python

# 函數工具
functools  # 內建於Python
operator  # 內建於Python

# 錯誤處理
traceback  # 內建於Python
inspect  # 內建於Python

# 數據類
dataclasses  # 內建於Python (Python 3.7+)

# 枚舉
enum  # 內建於Python

# 抽象基類
abc  # 內建於Python

# 上下文管理
contextlib  # 內建於Python

# 複製
copy  # 內建於Python

# 隨機數
random  # 內建於Python

# 字符串處理
string  # 內建於Python
textwrap  # 內建於Python

# 單元測試
unittest  # 內建於Python

# 模擬和打補丁
unittest.mock  # 內建於Python

# 臨時文件
tempfile  # 內建於Python

# 全局變量和配置
globals  # 內建函數
locals   # 內建函數

# 元編程
types  # 內建於Python
importlib  # 內建於Python

# 弱引用
weakref  # 內建於Python

# 二進制數據
struct  # 內建於Python
base64  # 內建於Python

# 編碼和解碼
codecs  # 內建於Python
locale  # 內建於Python

# 輸入輸出
io  # 內建於Python

# 子進程
subprocess  # 內建於Python

# 信號處理
signal  # 內建於Python

# 內存映射
mmap  # 內建於Python

# 配置解析器
configparser  # 內建於Python

# HTML和XML解析
html.parser  # 內建於Python
xml.etree.ElementTree  # 內建於Python

# URL處理
urllib.parse  # 內建於Python

# 郵件處理
email  # 內建於Python

# FTP客戶端
ftplib  # 內建於Python

# HTTP客戶端和服務器
http.client  # 內建於Python
http.server  # 內建於Python

# 選項解析
optparse  # 內建於Python (已棄用，推薦使用argparse)

# 命令行工具
cmd  # 內建於Python

# 調試器
pdb  # 內建於Python

# 性能分析
profile  # 內建於Python
cProfile  # 內建於Python

# 文檔測試
doctest  # 內建於Python

# 幫助系統
help  # 內建函數
pydoc  # 內建於Python

# 包管理
pkgutil  # 內建於Python
distutils  # 內建於Python

# 資源管理
resource  # 平台相關

# 加密
ssl  # 內建於Python

# 套接字服務器
socketserver  # 內建於Python

# 線程本地數據
threading.local  # 內建於Python

# 隊列
queue  # 內建於Python

# 事件調度
sched  # 內建於Python

# 堆隊列
heapq  # 內建於Python

# 雙端隊列
collections.deque  # 內建於Python

# 有序字典
collections.OrderedDict  # 內建於Python

# 計數器
collections.Counter  # 內建於Python

# 默認字典
collections.defaultdict  # 內建於Python

# 命名元組
collections.namedtuple  # 內建於Python

# 鏈映射
collections.ChainMap  # 內建於Python

# 用戶字典和列表
collections.UserDict  # 內建於Python
collections.UserList  # 內建於Python
collections.UserString  # 內建於Python